Remove volatile from _receive_buffer_tail/head
Making these two volatile does not make sense. In ISR context it has no
use, and outside ISR context they are only used in simple logic (no need
to reload values).

However, there could be an issue with the usage of the combination of these
variables. But "volatile" is not going to help to avoid problems. We may
need a cli/sti guard in some functions.