Use SoftwareSerial from Arduino 1.6
This version uses Sodaq_PcInt library to handle the PCINT interrupts. You
need a update of Sodaq_PcInt with the new enable / disable interrupt
functions.

Please notice the (experimental) -30 in ::recv() for the delay to get to
the middle of the start bit. This needs to be verified using the debug
mode and using a scope. This -30 is to compensate for the less efficient
overhead of Sodaq_PcInt to get into the recv() handler.